+++ oresoresores-1.18.2/common/src/main/java/ooo/item/RawCrappium.java
@@ -1,0 +1,13 @@
+package ooo.item;
+
+import dev.architectury.registry.registries.RegistrySupplier;
+import net.minecraft.world.item.Item;
+import ooo.OresOresOres;
+
+public class RawCrappium extends Item {
+    public static RegistrySupplier<Item> ITEM;
+
+    public RawCrappium() {
+        super(new Properties().tab(OresOresOres.TAB));
+    }
+}